Update unikraft related Dockerfiles
Unikraft supports vaccelrt, so we need to update the Dockerfiles so they
can reflect the changes. From now on we use vaccelrt for both the host and
the unikraft applications.

Moreover When we execute the run.sh script we should be able to specify which
image we want to classify and the number of iterations.

Also it is better to mount guest data directory when running the
container so we can easily add new images without having to build any
container again.

Signed-off-by: Charalampos Mainas <charalampos.mainas@gmail.com>